fix: move config and intercom out of GlobalContext
