+++ services/orchest-webserver/client/src/hooks/useHideIntercom.tsx
@@ -1,4 +1,4 @@
-import { useGlobalContext } from "@/contexts/GlobalContext";
+import { hasValue } from "@orchest/lib-utils";
 import React from "react";
 
 /** Hide Intercom in case the UI component is blocked by it */
@@ -6,8 +6,22 @@ export const useHideIntercom = (
   shouldHidIntercom: boolean,
   shouldAutoShow = true
 ) => {
-  const { hideIntercom, showIntercom } = useGlobalContext();
+  const setDisplayOfIntercom = React.useCallback((value: "block" | "none") => {
+    const intercomElement = document.querySelector(
+      ".intercom-lightweight-app"
+    ) as HTMLElement;
 
+    if (hasValue(intercomElement?.style?.display))
+      intercomElement.style.display = value;
+  }, []);
+
+  const hideIntercom = React.useCallback(() => {
+    setDisplayOfIntercom("none");
+  }, [setDisplayOfIntercom]);
+
+  const showIntercom = React.useCallback(() => {
+    setDisplayOfIntercom("block");
+  }, [setDisplayOfIntercom]);
   React.useEffect(() => {
     if (shouldHidIntercom) {
       hideIntercom();
